+++ homeassistant/components/homematic/__init__.py
@@ -18,7 +18,6 @@ from homeassistant.const import (
 from homeassistant.helpers import discovery
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity import Entity
-from homeassistant.loader import bind_hass
 
 REQUIREMENTS = ['pyhomematic==0.1.49']
 
@@ -243,78 +242,6 @@ SCHEMA_SERVICE_PUT_PARAMSET = vol.Schema({
     vol.Required(ATTR_PARAMSET_KEY): vol.All(cv.string, vol.Upper),
     vol.Required(ATTR_PARAMSET): dict,
 })
-
-
-@bind_hass
-def virtualkey(hass, address, channel, param, interface=None):
-    """Send virtual keypress to homematic controller."""
-    data = {
-        ATTR_ADDRESS: address,
-        ATTR_CHANNEL: channel,
-        ATTR_PARAM: param,
-        ATTR_INTERFACE: interface,
-    }
-
-    hass.services.call(DOMAIN, SERVICE_VIRTUALKEY, data)
-
-
-@bind_hass
-def set_variable_value(hass, entity_id, value):
-    """Change value of a Homematic system variable."""
-    data = {
-        ATTR_ENTITY_ID: entity_id,
-        ATTR_VALUE: value,
-    }
-
-    hass.services.call(DOMAIN, SERVICE_SET_VARIABLE_VALUE, data)
-
-
-@bind_hass
-def set_device_value(hass, address, channel, param, value, interface=None):
-    """Call setValue XML-RPC method of supplied interface."""
-    data = {
-        ATTR_ADDRESS: address,
-        ATTR_CHANNEL: channel,
-        ATTR_PARAM: param,
-        ATTR_VALUE: value,
-        ATTR_INTERFACE: interface,
-    }
-
-    hass.services.call(DOMAIN, SERVICE_SET_DEVICE_VALUE, data)
-
-
-@bind_hass
-def put_paramset(hass, interface, address, paramset_key, paramset):
-    """Call putParamset XML-RPC method of supplied interface."""
-    data = {
-        ATTR_INTERFACE: interface,
-        ATTR_ADDRESS: address,
-        ATTR_PARAMSET_KEY: paramset_key,
-        ATTR_PARAMSET: paramset,
-    }
-
-    hass.services.call(DOMAIN, SERVICE_PUT_PARAMSET, data)
-
-
-@bind_hass
-def set_install_mode(hass, interface, mode=None, time=None, address=None):
-    """Call setInstallMode XML-RPC method of supplied interface."""
-    data = {
-        key: value for key, value in (
-            (ATTR_INTERFACE, interface),
-            (ATTR_MODE, mode),
-            (ATTR_TIME, time),
-            (ATTR_ADDRESS, address)
-        ) if value
-    }
-
-    hass.services.call(DOMAIN, SERVICE_SET_INSTALL_MODE, data)
-
-
-@bind_hass
-def reconnect(hass):
-    """Reconnect to CCU/Homegear."""
-    hass.services.call(DOMAIN, SERVICE_RECONNECT, {})
 
 
 def setup(hass, config):
