+++ homeassistant/components/input_number.py
@@ -15,7 +15,6 @@ from homeassistant.const import (
 from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.entity_component import EntityComponent
 from homeassistant.helpers.restore_state import async_get_last_state
-from homeassistant.loader import bind_hass
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -80,31 +79,6 @@ CONFIG_SCHEMA = vol.Schema({
         }, _cv_input_number)
     })
 }, required=True, extra=vol.ALLOW_EXTRA)
-
-
-@bind_hass
-def set_value(hass, entity_id, value):
-    """Set input_number to value."""
-    hass.services.call(DOMAIN, SERVICE_SET_VALUE, {
-        ATTR_ENTITY_ID: entity_id,
-        ATTR_VALUE: value,
-    })
-
-
-@bind_hass
-def increment(hass, entity_id):
-    """Increment value of entity."""
-    hass.services.call(DOMAIN, SERVICE_INCREMENT, {
-        ATTR_ENTITY_ID: entity_id
-    })
-
-
-@bind_hass
-def decrement(hass, entity_id):
-    """Decrement value of entity."""
-    hass.services.call(DOMAIN, SERVICE_DECREMENT, {
-        ATTR_ENTITY_ID: entity_id
-    })
 
 
 @asyncio.coroutine
