+++ tests/components/image_processing/test_microsoft_face_identify.py
@@ -9,6 +9,7 @@ import homeassistant.components.microsoft_face as mf
 
 from tests.common import (
     get_test_home_assistant, assert_setup_component, load_fixture, mock_coro)
+from tests.components.image_processing import common
 
 
 class TestMicrosoftFaceIdentifySetup:
@@ -150,7 +151,7 @@ class TestMicrosoftFaceIdentify:
             text=load_fixture('microsoft_face_identify.json')
         )
 
-        ip.scan(self.hass, entity_id='image_processing.test_local')
+        common.scan(self.hass, entity_id='image_processing.test_local')
         self.hass.block_till_done()
 
         state = self.hass.states.get('image_processing.test_local')
