+++ tests/components/image_processing/test_openalpr_local.py
@@ -9,6 +9,7 @@ import homeassistant.components.image_processing as ip
 
 from tests.common import (
     get_test_home_assistant, assert_setup_component, load_fixture)
+from tests.components.image_processing import common
 
 
 @asyncio.coroutine
@@ -146,7 +147,7 @@ class TestOpenAlprLocal:
         """Set up and scan a picture and test plates from event."""
         aioclient_mock.get(self.url, content=b'image')
 
-        ip.scan(self.hass, entity_id='image_processing.test_local')
+        common.scan(self.hass, entity_id='image_processing.test_local')
         self.hass.block_till_done()
 
         state = self.hass.states.get('image_processing.test_local')
