+++ tests/components/test_input_number.py
@@ -4,11 +4,48 @@ import asyncio
 import unittest
 
 from homeassistant.core import CoreState, State, Context
-from homeassistant.setup import setup_component, async_setup_component
 from homeassistant.components.input_number import (
-    DOMAIN, set_value, increment, decrement)
+    ATTR_VALUE, DOMAIN, SERVICE_DECREMENT, SERVICE_INCREMENT,
+    SERVICE_SET_VALUE)
+from homeassistant.const import ATTR_ENTITY_ID
+from homeassistant.loader import bind_hass
+from homeassistant.setup import setup_component, async_setup_component
 
 from tests.common import get_test_home_assistant, mock_restore_cache
+
+
+@bind_hass
+def set_value(hass, entity_id, value):
+    """Set input_number to value.
+
+    This is a legacy helper method. Do not use it for new tests.
+    """
+    hass.services.call(DOMAIN, SERVICE_SET_VALUE, {
+        ATTR_ENTITY_ID: entity_id,
+        ATTR_VALUE: value,
+    })
+
+
+@bind_hass
+def increment(hass, entity_id):
+    """Increment value of entity.
+
+    This is a legacy helper method. Do not use it for new tests.
+    """
+    hass.services.call(DOMAIN, SERVICE_INCREMENT, {
+        ATTR_ENTITY_ID: entity_id
+    })
+
+
+@bind_hass
+def decrement(hass, entity_id):
+    """Decrement value of entity.
+
+    This is a legacy helper method. Do not use it for new tests.
+    """
+    hass.services.call(DOMAIN, SERVICE_DECREMENT, {
+        ATTR_ENTITY_ID: entity_id
+    })
 
 
 class TestInputNumber(unittest.TestCase):
